+++ core/src/fetched-record-buffer.ts
@@ -1,35 +1,29 @@
 import { FetchedRecord } from './fetched-record'
 
 export class FetchedRecordBuffer {
-  store: { [subject: string]: { [position: string]: FetchedRecord[] } }
+  store: { [subject: string]: { [transactionId: string]: FetchedRecord[] } }
 
   constructor() {
     this.store = {}
   }
 
-  static fromStore(store: { [subject: string]: { [position: string]: FetchedRecord[] } }) {
-    const buffer = new FetchedRecordBuffer()
-    buffer.store = store
-    return buffer
-  }
-
   addFetchedRecord(fetchedRecord: FetchedRecord) {
     const newBuffer: FetchedRecordBuffer = Object.assign(Object.create(this), this)
     const { subject, changeAttributes } = fetchedRecord
-    const position = changeAttributes.position.toString()
-    const existingFetchedRecords = this.store[subject]?.[position]
+    const transactionId = changeAttributes.transactionId.toString()
+    const existingFetchedRecords = this.store[subject]?.[transactionId]
 
     if (existingFetchedRecords) {
       newBuffer.store = {
         ...this.store,
-        [subject]: { ...this.store[subject], [position]: [...existingFetchedRecords, fetchedRecord] },
+        [subject]: { ...this.store[subject], [transactionId]: [...existingFetchedRecords, fetchedRecord] },
       }
       return newBuffer
     }
 
     newBuffer.store = {
       ...this.store,
-      [subject]: { ...(this.store[subject] || []), [position]: [fetchedRecord] },
+      [subject]: { ...(this.store[subject] || []), [transactionId]: [fetchedRecord] },
     }
     return newBuffer
   }
@@ -48,17 +42,21 @@ export class FetchedRecordBuffer {
   forEach(callback: (subject: string, fetchedRecords: FetchedRecord[]) => void) {
     Object.keys(this.store).forEach((subject) => {
       const fetchedRecords = Object.values(this.store[subject]).flat()
-      const sortedFetchedRecords = fetchedRecords.sort(
-        (a, b) =>
-          parseInt(a.changeAttributes.position.toString(), 10) - parseInt(b.changeAttributes.position.toString(), 10),
-      )
+
+      // Sort by transactionId, then by streamSequence
+      const sortedFetchedRecords = fetchedRecords.sort((a, b) => {
+        if (a.changeAttributes.transactionId !== b.changeAttributes.transactionId) {
+          return a.changeAttributes.transactionId - b.changeAttributes.transactionId
+        }
+        return a.streamSequence - b.streamSequence
+      })
 
       callback(subject, sortedFetchedRecords)
     })
   }
 
-  fetchedRecordsByPosition(subject: string, position: string) {
-    return this.store[subject]?.[position] || []
+  fetchedRecordsByTransactionId(subject: string, transactionId: string) {
+    return this.store[subject]?.[transactionId] || []
   }
 
   sizeBySubject(subject: string) {